refactor: Update v2 to v3 in validator block references
