+++ pkg/sentry/event/beacon/eth/v3/proposed_validator_block.go
@@ -1,4 +1,4 @@
-package event
+package v3
 
 import (
 	"context"
