+++ pkg/server/service/event-ingester/event/beacon/eth/v3/proposed_validator_block.go
@@ -1,4 +1,4 @@
-package v2
+package v3
 
 import (
 	"context"
